move structs into models package
