+++ master/pkg/model/task_config_policy.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/determined-ai/determined/master/pkg/schemas/configpolicy"
 	"github.com/determined-ai/determined/master/pkg/schemas/expconf"
 	"github.com/determined-ai/determined/proto/pkg/configpolicyv1"
 	"github.com/uptrace/bun"
@@ -32,18 +31,33 @@ type ExperimentTaskConfigPolicies struct {
 	LastUpdatedTime time.Time                 `bun:"last_updated_time,notnull"`
 	WorkloadType    WorkloadType              `bun:"workload_type,notnull"`
 	InvariantConfig *expconf.ExperimentConfig `bun:"invariant_config"`
-	Constraints     *configpolicy.Constraints `bun:"constraints"`
+	Constraints     Constraints               `bun:"constraints"`
 }
 
 // NTSCTaskConfigPolicy is the bun model of a task config policy.
 type NTSCTaskConfigPolicies struct {
 	bun.BaseModel   `bun:"table:task_config_policies"`
-	WorkspaceID     *int                      `bun:"workspace_id,unique"`
-	LastUpdatedBy   UserID                    `bun:"last_updated_by,notnull"`
-	LastUpdatedTime time.Time                 `bun:"last_updated_time,notnull"`
-	WorkloadType    WorkloadType              `bun:"workload_type,notnull"`
-	InvariantConfig *CommandConfig            `bun:"invariant_config"`
-	Constraints     *configpolicy.Constraints `bun:"constraints"`
+	WorkspaceID     *int           `bun:"workspace_id,unique"`
+	LastUpdatedBy   UserID         `bun:"last_updated_by,notnull"`
+	LastUpdatedTime time.Time      `bun:"last_updated_time,notnull"`
+	WorkloadType    WorkloadType   `bun:"workload_type,notnull"`
+	InvariantConfig *CommandConfig `bun:"invariant_config"`
+	Constraints     Constraints    `bun:"constraints"`
+}
+
+// ResourceConstraints are non-overridable resource constraints.
+// Submitted workloads that request resource quanities exceeding defined resource constraints in a
+// given scope are rejected.
+type ResourceConstraints struct {
+	MaxSlots *int `json:"max_slots"`
+}
+
+// Constraints are non-overridable workload constraints.
+// Submitted workloads whose config's respective field(s) exceed defined constraints within a given
+// scope are rejected.
+type Constraints struct {
+	ResourceConstraints *ResourceConstraints `json:"resources"`
+	PriorityLimit       *int                 `json:"priority_limit"`
 }
 
 // WorkloadTypeFromProto maps taskconfigpolicyv1.WorkloadType to WorkloadType.
